+++ sourcecode/OutlookRecipientConfirmationAddin/RecipientConfirmationWindow.cs
@@ -10,29 +10,22 @@ using System.Windows.Forms;
 
 namespace OutlookRecipientConfirmationAddin
 {
-    public partial class RecipientConfirmationWindow : Form
+    public partial class RecipientConfirmationWindow : RecipientCommonWindow
     {
-        public enum SendType { Mail, Meeting, Appointment, MeetingResp };
 
-        SendType _type;
-        List<String> toList;
-        List<String> ccList;
-        List<String> bccList;
+        Utility.SendType _type;
+        List<RecipientInformationDto> _recipientsList;
 
         public RecipientConfirmationWindow()
         {
             InitializeComponent();
-            Anchor = (AnchorStyles.Top | AnchorStyles.Bottom | AnchorStyles.Left | AnchorStyles.Right);
         }
 
-        public RecipientConfirmationWindow(SendType type, List<String> toList, List<String> ccList, List<String> bccList)
+        public RecipientConfirmationWindow(Utility.SendType type, List<RecipientInformationDto> recipients) : base(type, recipients)
         {
             InitializeComponent();
-
-            _type = type;
-            this.toList = toList;
-            this.ccList = ccList;
-            this.bccList = bccList;
+            this._type = type;
+            this._recipientsList = recipients;
         }
 
         /// <summary>
@@ -42,53 +35,8 @@ namespace OutlookRecipientConfirmationAddin
         /// <param name="e"></param>
         private void RecipientConfirmationWindow_Load(object sender, EventArgs e)
         {
-            string firstHeader = "", secondHeder = "", thirdHeader = "";
-            switch (_type)
-            {
-                case SendType.Mail:
-                    firstHeader = "To";
-                    secondHeder = "Cc";
-                    thirdHeader = "Bcc";
-                    break;
-
-                case SendType.Meeting:
-                    firstHeader = "参加者";
-                    secondHeder = "参加者(任意)";
-                    thirdHeader = "リソース";
-                    break;
-            }
-
-            textBox1.Text = string.Format("■------------ {0}: {1}件 ------------■\r\n", firstHeader, toList.Count());
-            foreach (var recipients in toList)
-            {
-                textBox1.Text += recipients + "\r\n";
-            }
-            textBox1.AppendText("\r\n");
-
-            textBox1.Text += string.Format("■------------ {0}: {1}件 ------------■\r\n", secondHeder, ccList.Count());
-            foreach (var recipients in ccList)
-            {
-                textBox1.Text += recipients + "\r\n";
-            }
-            textBox1.AppendText("\r\n");
-
-            textBox1.Text += string.Format("■------------ {0}: {1}件 ------------■\r\n", thirdHeader, bccList.Count());
-            foreach (var recipients in bccList)
-            {
-                textBox1.Text += recipients + "\r\n";
-            }
-
-            /// 読み取り専用、自動で折り返さない
-            textBox1.ReadOnly = true;
-            textBox1.WordWrap = false;
-
-            /// 必要な場合、垂直、水平両方のスクロールバーを表示
-            textBox1.ScrollBars = ScrollBars.Both;
-
-            /// アンカーを設定
-            textBox1.Anchor = (AnchorStyles.Top | AnchorStyles.Bottom | AnchorStyles.Left | AnchorStyles.Right);
-            OK.Anchor = (AnchorStyles.Bottom | AnchorStyles.Right);
-            Cancel.Anchor = (AnchorStyles.Bottom | AnchorStyles.Right);
+            /// baseクラスでテキストボックスの内容を作る
+            RecipientCommonWindow_format();
         }
 
         /// <summary>
