+++ sourcecode/OutlookRecipientConfirmationAddin/RecipientListRibbon.cs
@@ -16,8 +16,8 @@ namespace OutlookRecipientConfirmationAddin
     [ComVisible(true)]
     public class RecipientListRibbon : Office.IRibbonExtensibility
     {
-        private const string TANTOU = "担当";
-        private const string PROPTAG_URL = "http://schemas.microsoft.com/mapi/proptag/0x0C190102";
+        //private const string TANTOU = "担当";
+        //private const string PROPTAG_URL = "http://schemas.microsoft.com/mapi/proptag/0x0C190102";
 
         private Office.IRibbonUI ribbon;
 
@@ -92,8 +92,7 @@ namespace OutlookRecipientConfirmationAddin
             {
                 activeItem = Globals.ThisAddIn.Application.ActiveInspector().CurrentItem;
             }
-
-            /// 
+            
             ShowRecipientListWindow(activeItem);
         }
 
@@ -102,148 +101,38 @@ namespace OutlookRecipientConfirmationAddin
         /// </summary>
         private void ShowRecipientListWindow(object activeItem)
         {
-            Outlook.Recipients recipients = null;
-            Outlook.AddressEntry sender = null;
-            Outlook.ExchangeUser exchUser = null;
-            Outlook.PropertyAccessor propAccess = null;
+            /// Mailで初期化
+            Utility.SendType itemType = Utility.SendType.Mail;
             RecipientInformationDto senderInformation = null;
 
-
-            /// Mailで初期化
-            RecipientConfirmationWindow.SendType type = RecipientConfirmationWindow.SendType.Mail;
-
-            /// 表示しているのがMailItemの場合
-            if (activeItem is Outlook.MailItem)
-            {
-                Outlook.MailItem mail = (activeItem as Outlook.MailItem);
-
-                //受信者情報を取得
-                recipients = mail.Recipients;
-
-                ///送信元のアカウントのユーザーに対応するSenderプロパティを取得
-                sender = mail.Sender;
-                if (sender != null)
-                {
-                    Outlook.Recipient recResolve = Globals.ThisAddIn.Application.Session.CreateRecipient(sender.Address);
-                    exchUser = recResolve.AddressEntry.GetExchangeUser();
-                }
-            }
-            /// MeetingItemの場合
-            else if (activeItem is Outlook.MeetingItem)
-            {
-                Outlook.MeetingItem meeting = activeItem as Outlook.MeetingItem;
-
-                //会議招集の返信
-                if (meeting.MessageClass.Contains("IPM.Schedule.Meeting.Resp."))
-                {
-                    type = RecipientConfirmationWindow.SendType.MeetingResp;
-                }
-                //会議招集依頼など
-                else
-                {
-                    type = RecipientConfirmationWindow.SendType.Meeting;
-                }
-
-                //受信者情報を取得
-                recipients = meeting.Recipients;
-
-                //送信者のExchangeUserを取得
-                propAccess = meeting.PropertyAccessor;
-                exchUser = FindExchangeUser(propAccess, ref sender);
-            }
-            /// AppointmentItemの場合(送信前の会議系のメール)
-            else if (activeItem is Outlook.AppointmentItem)
-            {
-                Outlook.AppointmentItem appointment = activeItem as Outlook.AppointmentItem;
-                type = RecipientConfirmationWindow.SendType.Appointment;
-
-                //受信者情報を取得
-                recipients = appointment.Recipients;
-
-                //送信者のExchangeUserを取得
-                propAccess = appointment.PropertyAccessor;
-                exchUser = FindExchangeUser(propAccess, ref sender);
-            }
-
-            /// メールでも会議招集でもない場合、なにも起きない
-            if (recipients == null)
-            {
-                return;
-            }
-
-            /// 受信者の情報をリストに入れる
+            /// メールの宛先を取得
+            Utility utility = new Utility();
             List<Outlook.Recipient> recipientsList = new List<Outlook.Recipient>();
-            foreach (Outlook.Recipient recipient in recipients)
-            {
-                recipientsList.Add(recipient);
-            }
+            recipientsList = utility.getRecipients(activeItem, ref itemType, false);
+
+            ///// メールでも会議招集でもない場合、なにも起きない
+            /// 空なら、空の画面を表示するにする？
+            //if (recipientsList == null)
+            //{
+            //    return;
+            //}
 
             /// 検索し、受信者の宛先情報リストが戻ってくる
             SearchRecipient searchRecipient = new SearchRecipient();
             List<RecipientInformationDto> recipientList = searchRecipient.SearchContact(recipientsList);
 
-            /// 送信者のExchangeUserオブジェクトが取得できた場合
-            if (exchUser != null)
+            /// 送信者のExchangeUserオブジェクトを取得
+            senderInformation = utility.GetSenderInfomation(activeItem);
+            
+            /// 受信者の宛先情報のリストに、送信者の情報も追加する
+            if (senderInformation != null)
             {
-                senderInformation = SetSenderInformation(exchUser);
-                // 受信者の宛先情報のリストに、送信者の情報も追加する
-                recipientList.Add(senderInformation);
-            }
-            /// 送信者のExchangeUserが見つからなかった場合、表示名を表示
-            else if (sender != null)
-            {
-                senderInformation = new RecipientInformationDto(sender.Name, Outlook.OlMailRecipientType.olOriginator);
-                // 受信者の宛先情報のリストに、送信者の情報も追加する
                 recipientList.Add(senderInformation);
             }
 
             // 宛先リストの画面を表示する
-            RecipientListWindow recipientListWindow = new RecipientListWindow(type, recipientList);
+            RecipientListWindow recipientListWindow = new RecipientListWindow(itemType, recipientList);
             recipientListWindow.ShowDialog();
-        }
-
-        /// <summary>
-        /// 送信者のExchangeUserを取得する（MeetingItemとAppointmentItem用）        
-        /// <param name="propAccess"></param>
-        /// <returns></returns>
-        /// </summary>
-        private Outlook.ExchangeUser FindExchangeUser(Outlook.PropertyAccessor propAccess, ref Outlook.AddressEntry sender)
-        {
-            /// PropoertyAccessorで、送信者の情報を取得
-            string senderID = propAccess.BinaryToString(propAccess.GetProperty(PROPTAG_URL));
-            sender = Globals.ThisAddIn.Application.Session.GetAddressEntryFromID(senderID);
-
-            /// 送信者のメールアドレスから、ExchangeUserを見つける
-            Outlook.Recipient recResolve = Globals.ThisAddIn.Application.Session.CreateRecipient(sender.Address);
-            Outlook.ExchangeUser exchUser = recResolve.AddressEntry.GetExchangeUser();
-
-            return exchUser;
-        }
-
-        /// <summary>
-        /// 送信者のExchangeUserプロパティから、表示に必要な情報をセットする
-        /// </summary>
-        /// <param name="exchUser"></param>
-        /// <returns></returns>
-        private RecipientInformationDto SetSenderInformation(Outlook.ExchangeUser exchUser)
-        {
-            Outlook.ContactItem contactItem = null;
-
-            contactItem = Globals.ThisAddIn.Application.CreateItem(Outlook.OlItemType.olContactItem);
-            contactItem.FullName = exchUser.Name;
-            contactItem.CompanyName = exchUser.CompanyName;
-            contactItem.Department = exchUser.Department;
-
-            /// 表示する役職ならDtoに、違えば空文字を入れる
-            string jobTitle = exchUser.JobTitle;
-            if (TANTOU.Equals(contactItem.JobTitle) || contactItem.JobTitle == null)
-            {
-                jobTitle = "";
-            }
-
-            return new RecipientInformationDto(contactItem.FullName, contactItem.Department,
-                contactItem.CompanyName, jobTitle, Outlook.OlMailRecipientType.olOriginator);
-
         }
         #endregion
